refactor(iocuak): move binding utils to the binding folder
