+++ packages/iocuak/src/binding/utils/domain/bind.ts
